Allow inputs to not have form-control as class attr
We are using a theme built on bootstrap 3. However we have cases where the input class should be simply =col-xs-X for width control. When we set this on the form we still get form-control and so the input is wide and ignores the other class. It took awhile to find this where everything gets the form-control regardless of having a class already set. It seems to me that if you are setting the class you should set *all* classes required for what you are doing. This is a proof of concept change at the moment as I didn't touch the textarea/buttons or any other widget that uses the same syntax. Making this change locally keeps all other input fields properly styled and allows us to override the class attribute as I would expect. Our only other choice is to override the bootstrap_3_layout.html.twig if this type of change is unwanted. If it is wanted, then I can make the rest of the changes.